Move AccessorEntityRegistry into accessor package, rename, and put in correct json
