+++ src/main/java/io/github/minecraftcursedlegacy/impl/registry/RegistryImpl.java
@@ -1,9 +1,9 @@
 package io.github.minecraftcursedlegacy.impl.registry;
 
+import io.github.minecraftcursedlegacy.accessor.AccessorEntityRegistry;
 import io.github.minecraftcursedlegacy.api.registry.Id;
 import io.github.minecraftcursedlegacy.api.registry.Registry;
 import io.github.minecraftcursedlegacy.impl.Hacks;
-import io.github.minecraftcursedlegacy.mixin.MixinEntityRegistry;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.item.ItemType;
 import net.minecraft.item.TileItem;
@@ -109,7 +109,7 @@ public class RegistryImpl implements ModInitializer {
 	static {
 		//noinspection ResultOfMethodCallIgnored
 		Tile.BED.hashCode(); // make sure tiles are initialised
-		MixinEntityRegistry.getID_TO_CLASS(); // make sure entities are initialised
+		AccessorEntityRegistry.getID_TO_CLASS(); // make sure entities are initialised
 		ITEM_TYPE = new ItemTypeRegistry(new Id("api:item_type"));
 		TILE = new TileRegistry(new Id("api:tile"));
 		ENTITY_TYPE = new EntityTypeRegistry(new Id("api:entity_type"));
