+++ src/main/resources/api.mixins.json
@@ -4,7 +4,6 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "MixinBiome",
-    "MixinEntityRegistry",
     "MixinItemType",
     "MixinMcRegionLevelStorage",
     "MixinOldLevelStorage",
